refactor: разбиение utils
