+++ src/mock/film.js
@@ -1,19 +1,6 @@
-import dayjs from "dayjs";
 import {EMOJI} from "../const.js";
+import {getRandomInteger, generateDate} from "../utils/common.js";
 
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-const generateDate = () => {
-  const maxDaysGap = 365;
-  const daysGap = getRandomInteger(-maxDaysGap, 0);
-
-  return dayjs().add(daysGap, `day`).toDate();
-};
 
 const descriptions = [
   `Lorem ipsum dolor sit amet, consectetur adipiscing elit.`,
